fixed edit video list function
